==> 1_rockfish_taxonomic_assignment.Rmd <==
---
title: "using custom rockfish database to assign ASVs to taxa"
author: "Kimberly Ledger"
date: "2024-11-06"
output: html_document
---

```{r load-libraries}
library(tidyverse)
rename <- dplyr::rename
```

input blastn results from custom rockfish database
```{r}
rock <- read.table("/home/kimberly.ledger/rockfish_mb/data/blastn_tissue_mock_bt22_534_20250117.txt")  %>%
  filter(V4 > 100) %>% # make sure all retained matches are >100 bp
  select(-V5, -V6, -V7, -V8, -V9, -V10, -V11, -V12, -V13, -V14) %>%  #remove unnecessary columns
  group_by(V1, V2) %>% # group by the sequence key and the full taxonomy to reduce duplicate entries
  unique() %>% # doing that reduced the number of entries
  rename(qseqid=V1, accession_id=V2, perc_id=V3, length=V4) #%>% #rename headers
```

load the id's of accessions
```{r}
id <- read.csv("/home/kimberly.ledger/rockfish_mb/data/rkfish_ref_dbs/rockfish_reference_db_534_20250117.csv") %>%
  rename(accession_id = Name)

id %>% select(Organism) %>% unique()
```

join 
```{r}
rock_id <- rock %>%
  left_join(id) %>%
  rename(species = Organism)
```

check which what ASVS did not get an assignment
```{r}
temp <- rock_id %>% ungroup() %>% select(qseqid) %>% arrange() %>% unique() %>%
  separate(qseqid, into = c('x', 'num'), sep = "_")

observed <- as.integer(temp$num)
all <- 1:334

missing <- setdiff(all, observed)
missing
```

rank the hits by % identity
```{r}
seq_id_diff <- rock_id %>%
  select(-length) %>%
  group_by(qseqid, species) %>%
    mutate(seq_percID = max(perc_id)) %>%
    group_by(qseqid, species, seq_percID) %>%
  summarise(max(seq_percID)) %>% # take just the top hit for each taxon (for each sequence)
  select(-`max(seq_percID)`) %>%
  ungroup() %>%
  group_by(qseqid) %>%
       mutate(top_perc = max(seq_percID)) %>% # designate the highest % id for the best taxonomic hit in each sequence (in some, but not all cases, this is 100%)   %>%
      mutate(diff = top_perc - seq_percID) %>% # calculate the difference between the % identity of the top hit and each subsequent taxonomic hit
      arrange(diff)

seq_id_diff
```


since there are very few bp difference among species, retain matches within 0.5% of the top match and top match must be at least 95% 
```{r}
keepers <- seq_id_diff %>%
  filter(diff < 0.5) %>%
  filter(top_perc > 95)
```


now determine whether ASVs should be assigned to genus, family, order, etc. 
```{r}
singletons <- keepers %>%
  select(qseqid) %>%
  tally() %>%
  filter(n == 1)

# these are the seqs that now have only a single match
singleton_df <- singletons %>%
  left_join(keepers) %>%
  select(-n) %>%
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species) #%>%
  #select(qseqid, taxon, taxonomic_level)

## Genus-level matches
# remove the singletons from the bigger df 
single_genus <- keepers %>%
  anti_join(singleton_df)%>% # at best, these should be genus-level matches
  left_join(., keepers) %>%
  arrange(qseqid) %>%
  mutate(taxonomic_level = "genus") %>%
  mutate(taxon = "Sebastes") #%>%
  #select(qseqid, taxon, taxonomic_level) %>%
  #unique
```

change genus level ASV assignments so that they represent the species make-up 
- i.e. Sebastes_ev = Sebastes emphaeus or Sebastes variegatus
```{r}
single_genus <- single_genus %>%
  group_by(qseqid) %>%
  mutate(taxon = case_when(
    any(species == "Sebastes ciliatus") & any(species == "Sebastes ciliatus_variabilis") ~ "Sebastes ciliatus_variabilis",
    TRUE ~ taxon
  )) %>%
  ungroup()

single_genus <- single_genus %>%
  group_by(qseqid) %>%
  mutate(taxon = case_when(
    any(species == "Sebastes variabilis") & any(species == "Sebastes ciliatus_variabilis") ~ "Sebastes ciliatus_variabilis",
    TRUE ~ taxon
  )) %>%
  ungroup()

single_genus <- single_genus %>%
  group_by(qseqid) %>%
  mutate(taxon = case_when(
    any(species == "Sebastes variegatus") & any(species == "Sebastes wilsoni") ~ "Sebastes vw",
    TRUE ~ taxon
  )) %>%
  ungroup()

single_genus <- single_genus %>%
  group_by(qseqid) %>%
  mutate(taxon = case_when(
    any(species == "Sebastes emphaeus") & any(species == "Sebastes variegatus") ~ "Sebastes ev",
    TRUE ~ taxon
  )) %>%
  ungroup()

single_genus <- single_genus %>%
  group_by(qseqid) %>%
  mutate(taxon = case_when(
    any(species == "Sebastes diaconus") & any(species == "Sebastes entomelas") ~ "Sebastes de",
    TRUE ~ taxon
  )) %>%
  ungroup()

single_genus <- single_genus %>%
  group_by(qseqid) %>%
  mutate(taxon = case_when(
    any(species == "Sebastes polyspinis") & any(species == "Sebastes ciliatus_variabilis") ~ "Sebastes ciliatus_variabilis_polyspinus",
    TRUE ~ taxon
  )) %>%
  ungroup()
```

bind species and genus ids
```{r}
rockfish_df <- singleton_df %>%
  bind_rows(single_genus)
```

collapse to keep one taxon assignment per ASV
```{r}
rockfish_df_collapsed <- rockfish_df  %>%
   select(qseqid, taxon) %>%
   unique()
```

```{r}
#write.csv(rockfish_df_collapsed, "/home/kimberly.ledger/rockfish_mb/data/taxonomy_custom534_20250117.csv")
```
