==> 3C_qm_mock_figures.Rmd <==
---
title: "quantitative metabarcoding figures for rockfish mock communities"
author: "Kimberly Ledger"
date: "2024-12-27"
output: html_document
---

load libraries
```{r}
library(tidyverse)
library(data.table)
library(gridExtra)
library(ggsci)
```

# Read in the posteriors for mock communities
```{r}
#load("/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/raw_20241227_nearshoreeven.Rdata")
#raw <- Output

load("/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/mock_20250117_allmocks.Rdata")
mock1 <- Output
```

```{r}
#########################################################
# Mock1
#########################################################
# summarize raw estimates from reads for each species.
mock1.raw <- mock1$env %>% group_by(community,Cycles,tech_rep) %>%
      mutate(sum.ng = sum(start_conc_ng),
             true.prop = start_conc_ng / sum.ng) %>%
      ungroup() %>%
      group_by(Species,community,Cycles,true.prop) %>%
  summarize(simple.Mean=mean(propReads),
            simple.N = length(tech_rep)) %>%
  replace_na(list(raw.Mean=0,raw.SD=0,raw.SE=0))

# extract predicted proportions from the posterior
COM <- data.frame(community = levels(mock1$env$community %>% as.factor()))
COM$comm_idx <- 1:nrow(COM)
SP  <- mock1$env %>% distinct(Species,sp_idx) %>% as.data.frame()

# These are the predicted intercepts for the posteriors
beta_posterior <- mock1$stanMod_summary[["int_samp_small"]][, c(1,4:8)]
colnames(beta_posterior) <- paste0("mock1.",substr(colnames(beta_posterior),1,nchar(colnames(beta_posterior))-1))
colnames(beta_posterior)[1] <- "mock1.mean"
beta_posterior <- as.data.frame(beta_posterior)

mock1.post <-expand.grid(comm_idx = COM$comm_idx,sp_idx =SP$sp_idx) %>% 
    arrange(comm_idx,sp_idx) %>% 
    left_join(.,COM) %>% 
    left_join(.,SP) %>% 
    bind_cols(.,beta_posterior)

# Combine the raw estimates and posterior estimates
mock1.all <- full_join(mock1.raw,mock1.post)
```

```{r}
###############################################################333
#### Pull out estimates of alpha, convert to CLR                        
###############################################################333

p_space_mock1 <- (exp(mock1$pars$alpha) / rowSums(exp(mock1$pars$alpha))) %>% as.data.frame()

clr_alpha_list_mock <- compositions::clr(p_space_mock1)

clr_alpha <- clr_alpha_list_mock %>% as.data.frame()

mock.taxon   <- data.frame(Species=mock1.post %>% distinct(Species) %>% pull(Species))

colnames(clr_alpha) <- mock.taxon$Species                       

clr_alpha_sum <- clr_alpha %>% 
                    pivot_longer( .,
                          cols = colnames(clr_alpha),
                          names_to="Species",values_to="val") %>%
                    group_by(Species) %>%
                    summarize(Mean = mean(val),
                        SD=sd(val),
                        q.025 = quantile(val,probs=0.025),
                        q.05 = quantile(val,probs=0.05),
                        q.25 = quantile(val,probs=0.25),
                        q.75 = quantile(val,probs=0.75),
                        q.95 = quantile(val,probs=0.95),
                        q.975 = quantile(val,probs=0.975))    

clr_alpha_sum <- clr_alpha_sum %>% arrange(Mean)

# get rid of reference species denotation
clr_alpha_sum <- clr_alpha_sum %>%
                      mutate(SP= ifelse(grepl("zRefSpecies_",Species),
                                       substr(Species,13,nchar(Species)),
                                       as.character(Species))) %>%
  mutate(SP = sub("^(\\w)\\w*\\s(\\w+)", "\\1. \\2", SP))

clr_alpha_sum$SP <- as.factor(clr_alpha_sum$SP)


## make a table for species colors 
species_colors_table <- data.frame(name = rev(clr_alpha_sum$SP[order(clr_alpha_sum$Mean)]), color = c(
  "#77ab31", "#f77189", "#6e9bf4", "#f45cf2", "#ae9d31", "#2bbab8", 
  "#7f7f7f", "#1f77b4", "#f5f056", "#f7754f",#"#38a9c5", 
  "#f5a056", "#97a431", "#a48cf4", "#3aa5df", "#8c564b", "#f66bad",  "#c7c7c7", "#33b07a" #"#37abb4", 
))

clr_alpha_sum$SP <- factor(clr_alpha_sum$SP, levels = rev(clr_alpha_sum$SP[order(clr_alpha_sum$Mean)]))

p_clr <-  ggplot(clr_alpha_sum) +
    geom_errorbarh(aes(xmin=q.25,xmax=q.75,y=SP),size=2,height=0) +
    geom_errorbarh(aes(xmin=q.025,xmax=q.975,y=SP),size=0.8,height=0) +
    geom_point(aes(x=Mean,y=SP,fill=SP),size=3,shape=21) +
    geom_vline(xintercept=0,linetype="dashed") +
    scale_fill_manual(values=species_colors_table$color) +
    scale_x_continuous("Amplification Efficiency (CLR)") +
    scale_y_discrete(NULL) + #, limits = rev(levels(clr_alpha_sum$taxon))) +
    theme_bw() +
    theme(legend.position = "right",
          legend.text = element_text(size = 8),
          axis.text.y = element_blank(),
          axis.ticks.y = element_blank(),
          legend.title = element_blank(),
          legend.key.height = unit(0.75, "lines")) + # Decrease vertical spacing) +
  guides(fill = guide_legend(reverse = TRUE))

p_clr
```


```{r}
#ggsave(p_clr, filename = "/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/figures/allmocks_clr_ordered.png", width = 8, height = 5, units = "in")
```


make a plot for publication using  - all_even community expects & observed proportions 
```{r}
mockcom_data <- read.csv("/home/kimberly.ledger/rockfish_mb/outputs/rf_mock_20250117.csv") %>%
  group_by(community, tech_rep) %>%
  mutate(community_total_conc = sum(start_conc_ng),
         Perc_Community = start_conc_ng/community_total_conc)

library(ggsci)

df1 <- mockcom_data %>%
  filter(community == "all_even") %>%
  group_by(tech_rep) %>%
  mutate(total_reads = sum(nReads)) %>%
  mutate(prop_reads = nReads/total_reads) %>%
  ungroup() %>%
  group_by(taxon) %>%
  summarise(proportion = mean(prop_reads)) %>%
  mutate(status = "observed")

df2 <- mockcom_data %>%
  ungroup() %>%
  filter(community == "all_even") %>%
  filter(tech_rep == "A") %>%
  select(taxon, Perc_Community) %>%
  mutate(status = "expected") %>%
  rename(proportion = Perc_Community)

df_mock <- df1 %>%
  bind_rows(df2) %>%
  mutate(taxon = sub("^(\\w)\\w*\\s(\\w+)", "\\1. \\2", taxon))

species_colors <- c(
  "#77ab31", "#f77189", "#6e9bf4", "#f45cf2", "#ae9d31", "#cc7af4", 
  "#31b33e", "#f565cc", "#36ada4", "#97a431", "#f7754f",#"#38a9c5", 
  "#35ae93", "#c39532", "#3aa5df", "#a48cf4", "#f66bad",  "#dc8932", "#33b07a" #"#37abb4", 
)

df_mock$taxon <- as.factor(df_mock$taxon)

df_mock$taxon <- factor(df_mock$taxon, levels = clr_alpha_sum$SP)

alle <- df_mock %>%
  ggplot(aes(x=status, y=proportion, fill=taxon)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  scale_fill_manual(values = rev(species_colors)) +
  labs(
    y = "proportion",
    x = "") + 
  theme(#axis.text.x = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )

alle

temp <- df_mock %>%
  pivot_wider(names_from = "status", values_from = "proportion") %>%
  mutate(diff = observed - expected) %>%
  arrange(abs(diff))
```

```{r}
#ggsave(alle, filename = "/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/figures/alleven_initial_v_observed_ordered.png", width = 8, height = 5, units = "in")
```

```{r}
df1 <- mockcom_data %>%
  filter(community == "offshore_even") %>%
  filter(nReads > 10) %>%
  group_by(tech_rep) %>%
  mutate(total_reads = sum(nReads)) %>%
  mutate(prop_reads = nReads/total_reads) %>%
  ungroup() %>%
  group_by(taxon) %>%
  summarise(proportion = mean(prop_reads)) %>%
  mutate(status = "observed")

df2 <- mockcom_data %>%
  ungroup() %>%
  filter(community == "offshore_even") %>%
  filter(tech_rep == "A") %>%
  select(taxon, Perc_Community) %>%
  mutate(status = "expected") %>%
  rename(proportion = Perc_Community) %>%
  filter(proportion > 0)

df_mock <- df1 %>%
  bind_rows(df2) %>%
  mutate(taxon = sub("^(\\w)\\w*\\s(\\w+)", "\\1. \\2", taxon))

species_colors_os <- species_colors_table %>%
  filter(name %in% df_mock$taxon)

df_mock$taxon <- as.factor(df_mock$taxon)

df_mock$taxon <- factor(df_mock$taxon, levels = clr_alpha_sum$SP)

os_even <- df_mock %>%
  ggplot(aes(x=status, y=proportion, fill=taxon)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  scale_fill_manual(values = rev(species_colors_os$color)) +
  labs(
    y = "proportion",
    x = "") + 
  theme(#axis.text.x = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )

os_even
```

```{r}
df1 <- mockcom_data %>%
  filter(community == "nearshore_even") %>%
  filter(nReads > 30) %>%
  group_by(tech_rep) %>%
  mutate(total_reads = sum(nReads)) %>%
  mutate(prop_reads = nReads/total_reads) %>%
  ungroup() %>%
  group_by(taxon) %>%
  summarise(proportion = mean(prop_reads)) %>%
  mutate(status = "observed")

df2 <- mockcom_data %>%
  ungroup() %>%
  filter(community == "nearshore_even") %>%
  filter(tech_rep == "A") %>%
  select(taxon, Perc_Community) %>%
  mutate(status = "expected") %>%
  rename(proportion = Perc_Community) %>%
  filter(proportion > 0)

df_mock <- df1 %>%
  bind_rows(df2) %>%
  mutate(taxon = sub("^(\\w)\\w*\\s(\\w+)", "\\1. \\2", taxon))

species_colors_ns <- species_colors_table %>%
  filter(name %in% df_mock$taxon)

df_mock$taxon <- as.factor(df_mock$taxon) 

df_mock$taxon <- factor(df_mock$taxon, levels = clr_alpha_sum$SP)

ns_even <- df_mock %>%
  ggplot(aes(x=status, y=proportion, fill=taxon)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  scale_fill_manual(values = rev(species_colors_ns$color)) +
  labs(
    y = "proportion",
    x = "") + 
  theme(#axis.text.x = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )

ns_even
```


```{r}
library(patchwork)

mock_plot <- (os_even | ns_even) /  p_clr + plot_annotation(tag_levels = "A") + plot_layout(heights = c(1.5, 2))
mock_plot
```

```{r}
#ggsave(mock_plot, filename = "/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/figures/mock_fig_off_near_tall.png", width = 7, height = 7, units = "in")
```


```{r}
df1 <- mockcom_data %>%
  filter(community == "offshore_skew") %>%
  filter(nReads > 30) %>%
  group_by(tech_rep) %>%
  mutate(total_reads = sum(nReads)) %>%
  mutate(prop_reads = nReads/total_reads) %>%
  ungroup() %>%
  group_by(taxon) %>%
  summarise(proportion = mean(prop_reads)) %>%
  mutate(status = "observed")

df2 <- mockcom_data %>%
  ungroup() %>%
  filter(community == "offshore_skew") %>%
  filter(tech_rep == "A") %>%
  select(taxon, Perc_Community) %>%
  mutate(status = "expected") %>%
  rename(proportion = Perc_Community) %>%
  filter(proportion > 0)

df_mock <- df1 %>%
  bind_rows(df2) %>%
  mutate(taxon = sub("^(\\w)\\w*\\s(\\w+)", "\\1. \\2", taxon))

species_colors_ns <- species_colors_table %>%
  filter(name %in% df_mock$taxon)

df_mock$taxon <- as.factor(df_mock$taxon) 

df_mock$taxon <- factor(df_mock$taxon, levels = clr_alpha_sum$SP)

os_skew <- df_mock %>%
  ggplot(aes(x=status, y=proportion, fill=taxon)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  scale_fill_manual(values = rev(species_colors_ns$color)) +
  labs(
    y = "proportion",
    x = "") + 
  theme(#axis.text.x = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )

os_skew

temp <- df_mock %>%
  pivot_wider(names_from = "status", values_from = "proportion") %>%
  mutate(diff = observed - expected) %>%
  arrange(abs(diff))
```


```{r}
library(patchwork)

sup_plot <- alle + os_skew + plot_annotation(tag_levels = "A")
sup_plot
```

```{r}
ggsave(sup_plot, filename = "/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/figures/sup_fig_initial_v_observed_ordered.png", width = 8, height = 4, units = "in")
```


```{r}
alpha <- p_space_mock1 %>% as.data.frame()

mock.taxon   <- data.frame(Species=mock1.post %>% distinct(Species) %>% pull(Species))

colnames(alpha) <- mock.taxon$Species                       

alpha_sum <- alpha %>% 
                    pivot_longer( .,
                          cols = colnames(alpha),
                          names_to="Species",values_to="val") %>%
                    group_by(Species) %>%
                    summarize(Mean = mean(val),
                        SD=sd(val),
                        q.025 = quantile(val,probs=0.025),
                        q.05 = quantile(val,probs=0.05),
                        q.25 = quantile(val,probs=0.25),
                        q.75 = quantile(val,probs=0.75),
                        q.95 = quantile(val,probs=0.95),
                        q.975 = quantile(val,probs=0.975))    

alpha_sum <- alpha_sum %>% arrange(Mean)

# get rid of reference species denotation
alpha_sum <- alpha_sum %>%
                      mutate(SP= ifelse(grepl("zRefSpecies_",Species),
                                       substr(Species,13,nchar(Species)),
                                       as.character(Species))) %>%
  mutate(SP = sub("^(\\w)\\w*\\s(\\w+)", "\\1. \\2", SP))

alpha_sum$SP <- as.factor(alpha_sum$SP)
```

read in table with primer-template mismatches
```{r}
mm <- read.csv("/home/kimberly.ledger/rockfish_mb/data/primer_mismatch_table - Sheet1.csv") %>%
  mutate(SP = Species) %>%
  mutate(SP = sub("^(\\w)\\w*\\s(\\w+)", "\\1. \\2", SP))
```

```{r}
alpha_mm <- alpha_sum %>%
  left_join(mm, by = "SP")
```

plot
```{r}
ggplot(alpha_mm, aes(x = HV.total, y = Mean)) +
  geom_point() +
  labs(y = "Mean Amplification Efficiency",
       x = "Total Number of Mismatches") +
  theme_minimal()
```
