==> 3D_qm_field.Rmd <==
---
title: "Quantitative metabarcoding of AK Rockfish field samples"
author:"Kimberly Ledger"
date: "2025-06-17"
output: github_document
---

calculate bias-corrected mb read proportions for Aleutian Is BT data 

libraries
```{r}
library(tidyverse)
library(compositions)
  select <- dplyr::select  #eliminate conflict w MASS::select vs. dplyr::select
  rename <- dplyr::rename
library(rstan)
```

load field data -- 
- this is output from "5_aquaria_bt22.Rmd"
```{r}
field_table_w_meta <- read.csv("/home/kimberly.ledger/rockfish_mb/outputs/field_quant/raw_field_data.csv") %>%
  select(!X)
```

for Sebastes ciliatus_variabilis and Sebastes ciliatus_variabilis_polyspinus, i'll use S. ciliatus in the mock commmunity to adjust
- rename in field data 
```{r}
field_table_w_meta <- field_table_w_meta %>%
  mutate(taxon = ifelse(taxon == "Sebastes ciliatus_variabilis", "Sebastes ciliatus", taxon)) %>%
  mutate(taxon = ifelse(taxon == "Sebastes ciliatus_variabilis_polyspinus", "Sebastes ciliatus", taxon)) 
```

#REQUIRED FORMAT
#environmental (i.e., unknown) samples with the following column names:
  #community  -- the unique biological sample from which a set of sequences derived
  #Species   -- the biological species to which reads are assigned
  #nReads -- the number of reads assigned to that species in that community
  #tech_rep -- index of technical replicate, within community (e.g., if each community is sequenced 3 times, tech_rep will vary between 1 and 3)
  #Cycles -- number of PCR cycles the reaction underwent
  
```{r}
#format field data
env <- field_table_w_meta %>% 
  select(extraction_ID, taxon, tot_reads) %>%
  rename(community = extraction_ID) %>%
  rename(Species = taxon) %>%
  rename(nReads = tot_reads) %>%
  mutate(Cycles = 35) %>%
  mutate(tech_rep = "A") %>%
  mutate(start_conc_ng = NA)
  
env <- env %>%  
  group_by(community, tech_rep) %>% 
  mutate(propReads = nReads/sum(nReads), #calculate proportion of reads
          totReads = sum(nReads)) %>%  #calculate total reads for community
  group_by(Species) %>% 
  mutate(totalSpeciesReads = sum(nReads)) %>%  
  add_tally(nReads > 0, name = "totalOccurrences") %>% 
  filter(totalSpeciesReads > 0)
         
#assign most common species to be the reference species
mostCommon <- env %>% 
    group_by(Species) %>% 
    tally(nReads > 0) %>%
    arrange(desc(n)) %>% 
    head(1) %>% 
    pull(Species)
env$Species[env$Species == mostCommon] <- paste0("zRefSpecies_", mostCommon) #make the most common species the reference species
env <- env %>% 
    arrange(Species, community)
```

load mock community data
```{r}
mockcom_data <- read.csv("/home/kimberly.ledger/rockfish_mb/outputs/rf_mock_20250117.csv") %>%
  mutate(tech_rep = as.factor(tech_rep),   ## i changed this from as.numeric to as.factor because of my rep format
         Cycles = as.numeric(Cycles)) %>%
  filter(start_conc_ng > 0) %>%   ## remove any reads belonging to species NOT in the mock community - some of these could be misid'd asvs - should check
  mutate(nReads = as.integer(nReads))
```

unique taxa?
```{r}
mockcom_data %>%
  select(taxon) %>%
  unique() %>%
  arrange(taxon)
```

group species by mock community and calculate total concentration
```{r}
sp.by.comm <- mockcom_data %>% filter(Cycles==35) %>% 
  group_by(species=taxon,community) %>%
  summarize(conc = sum(start_conc_ng)) %>%
  mutate(conc2 = ifelse(conc>0,1,0)) %>%
  arrange(species,community) %>% group_by(community) %>%
  mutate(tot_conc = sum(conc),prop=conc/tot_conc)
```


#REQUIRED FORMAT
mock community (i.e., known) samples with the following column names:
  #community  -- the unique biological sample from which a set of sequences derived
  #Species   -- the biological species to which reads are assigned
  #nReads -- the number of reads assigned to that species in that community
  #tech_rep -- index of technical replicate, within community (e.g., if each community is sequenced 3 times, tech_rep will vary between 1 and 3)
  #Cycles -- number of PCR cycles the reaction underwent
  #start_conc_ng -- starting concentration of each species in each community, in nanograms

```{r}
#set up mock community samples
mc <- mockcom_data %>% 
  #filter(community %in% c("all_even")) %>%
    rename("Species" = "taxon")  #mock comm samples
mc$Species[mc$Species == mostCommon] <- paste0("zRefSpecies_", mostCommon)   #make the most common species the reference species
  
# Filter so that you only keep species in the environment samples that are in the mock community.
# It is ok to include species that are only in the mock community.
env <- env %>%
    filter(Species %in% mc$Species)%>% #limit to species occurring in mock community dataset
    arrange(Species, community)  
  
#double check
sum(!mc$Species %in% unique(env$Species)) # This can be non-zero
sum(!env$Species %in% unique(mc$Species)) # this had better be zero.
```

```{r}
# Make a single species list:
  sp.list   <- data.frame(Species = sort(unique(mc$Species)) ) %>% mutate(sp_idx =1:length(Species))
  N_species <- nrow(sp.list)
    
  comm.mock.list <- mc %>% group_by(community, tech_rep,Cycles) %>% summarize(n=length(tech_rep)) %>%
                        ungroup() %>% mutate(id=1:length(n))
  comm.env.list   <- env %>% group_by(community, tech_rep,Cycles) %>% summarize(n=length(tech_rep)) %>%
                        ungroup() %>% mutate(id=1:length(n))
    
    #make a list of species that are in mock community but not environment, 
    # expand grid to make it so the the environmental samples get padded with all the
    # missing species for all of the communities and all tech replicates.
    
    sp.comm.mc  <- expand_grid(Species = sp.list$Species, id = comm.mock.list$id) %>% 
                          left_join(.,sp.list %>% select(Species,sp_idx)) %>%
                          left_join(.,comm.mock.list %>% select(community,tech_rep,Cycles,id) ) %>% select(-id)
    sp.comm.env <- expand_grid(Species = sp.list$Species, id = comm.env.list$id) %>% 
                          left_join(.,sp.list %>% select(Species,sp_idx)) %>%
                          left_join(.,comm.env.list %>% select(community,tech_rep,Cycles,id) ) %>% select(-id)

    

    #convert to matrices
    # merge in species and indices first to make pivoting more efficient.
    
    mc  <- left_join(sp.comm.mc,mc) %>%   
              mutate(nReads = ifelse(is.na(nReads),0,nReads),
                  start_conc_ng = ifelse(is.na(start_conc_ng),0,start_conc_ng)) 
    env <- left_join(sp.comm.env,env) %>%
              mutate(nReads = ifelse(is.na(nReads),0,nReads),
                  start_conc_ng = ifelse(is.na(start_conc_ng),0,start_conc_ng))
    
    sample_data <- env %>% 
      ungroup() %>% 
      dplyr::select(community, sp_idx, nReads, tech_rep, Cycles) %>% 
      arrange(sp_idx) %>% 
      pivot_wider(names_from = "sp_idx", values_from = "nReads", values_fill = 0) 
    
    sample_data_small <- sample_data %>% filter(tech_rep== "A")
    
    mock_data <- mc %>% 
      ungroup() %>% 
      dplyr::select(community, sp_idx, nReads,tech_rep, Cycles) %>% 
      arrange(sp_idx) %>% 
      pivot_wider(names_from = "sp_idx", values_from = "nReads", values_fill = 0)

    mock_data_small <- mock_data %>% filter(tech_rep== "A")
```


```{r}
#proportions
p_mock <- mc %>% 
  select(community, tech_rep, sp_idx, start_conc_ng, Cycles) %>% 
  arrange(sp_idx) %>% 
  group_by(community, tech_rep, Cycles) %>% 
  mutate(prop_conc = start_conc_ng/sum(start_conc_ng)) %>% 
  select(-start_conc_ng) %>% #, -Species) %>% 
  pivot_wider(names_from = "sp_idx", values_from = "prop_conc", values_fill = 0) %>% 
  ungroup() %>% 
  arrange(community)
  #select(-community)

p_mock_small <- mc %>%
  filter(tech_rep == "A") %>%
  select(community, sp_idx, start_conc_ng, Cycles) %>%
  arrange(sp_idx) %>%
  group_by(community) %>%
  mutate(prop_conc = start_conc_ng/sum(start_conc_ng)) %>%
  select(-start_conc_ng) %>%  # -Species) %>%
  pivot_wider(names_from = "sp_idx", values_from = "prop_conc", values_fill = 0) %>%
  ungroup() %>%
  arrange(community)
  #select(-community, -Cycles)

  #calculate additive log ratios 
  alr_mock_true_prop <- p_mock[,4:(ncol(p_mock)-1)]*0
  #i=1
  for(i in 1:nrow(p_mock)){
    alr_mock_true_prop[i,] <- compositions::alr(p_mock[i,4:(ncol(p_mock))] + 1e-10)
  }
  alr_mock_true_prop[,N_species] <- 0 #adding explicit reference species column
  
  alr_mock_true_prop_small <- p_mock_small[,3:(ncol(p_mock_small)-1)]*0
  for(i in 1:nrow(p_mock_small)){
    alr_mock_true_prop_small[i,] <- compositions::alr(p_mock_small[i,3:(ncol(p_mock_small))] + 1e-10)
  }
  alr_mock_true_prop_small[,N_species] <- 0
```


#DESIGN MATRICES
  # mock communities first
  # species compositions (betas)
  # use mock_data  
```{r}
  N_pcr_mock <- mock_data$Cycles
  
  if(length(unique(mock_data$community))==1){
    formula_b <- Cycles ~ 1  # what is on the left side of the equation doesn't matter.
  } else {
    formula_b <- Cycles ~ community # what is on the left side of the equation doesn't matter.
  }
  model_frame <- model.frame(formula_b, mock_data)
  model_matrix_b_mock <- model.matrix(formula_b, model_frame)
  
  #formula_b <- obs_sp_1 ~ community
  model_frame <- model.frame(formula_b, mock_data_small)
  model_matrix_b_mock_small <- model.matrix(formula_b, model_frame)
  
  # efficiencies (alphas)
  formula_a <- community ~ Cycles -1
  model_frame <- model.frame(formula_a, mock_data)
  model_vector_a_mock <- model.matrix(formula_a, model_frame) %>% as.numeric()
  model_frame <- model.frame(formula_a, mock_data_small)
  model_vector_a_mock_small <- model.matrix(formula_a, model_frame) %>% as.numeric()
  
  N_obs_mock_small <- nrow(model_matrix_b_mock_small)
  N_obs_mock       <- nrow(mock_data)
  N_b_mock_col     <- ncol(model_matrix_b_mock)  
  
  # unknown communities second
  # species compositions (betas)

  # use sample_data
  
  N_pcr_samp <- sample_data$Cycles
  
  if(length(unique(sample_data$community))==1){
    formula_b <- Cycles ~ 1  
  } else {
    formula_b <- Cycles ~ community
  }
  model_frame <- model.frame(formula_b, sample_data)
  model_matrix_b_samp <- model.matrix(formula_b, model_frame)
  
  #formula_b <- obs_sp_1 ~ community 
  #p_samp_all$site <- as.factor(p_samp_all$site)
  model_frame <- model.frame(formula_b, sample_data_small)
  model_matrix_b_samp_small <- model.matrix(formula_b, model_frame)
  
  # efficiencies (alpha)
  formula_a <- community ~ Cycles -1
  model_frame <- model.frame(formula_a, sample_data)
  model_vector_a_samp <- model.matrix(formula_a, model_frame) %>% as.numeric()
  model_frame <- model.frame(formula_a, sample_data_small)
  model_vector_a_samp_small <- model.matrix(formula_a, model_frame) %>% as.numeric()
  
  #counters 
  N_obs_samp_small <- nrow(model_matrix_b_samp_small)
  N_obs_samp <- nrow(sample_data)
  N_b_samp_col <- ncol(model_matrix_b_samp)               
  
```

set up for stan model
```{r}
stan_data <- list(
  N_species = N_species,   # Number of species in data
  N_obs_samp = N_obs_samp, # Number of observed samples 
  N_obs_mock = N_obs_mock, # Number of observed mock samples
  N_obs_samp_small = N_obs_samp_small, # Number of observed samples 
  N_obs_mock_small = N_obs_mock_small, # Number of observed mock samples
  
  # Observed data of community matrices
  sample_data = sample_data %>% select(-community,-Cycles,-tech_rep),
  mock_data   = mock_data  %>% select(-community,-Cycles,-tech_rep),
  
  # True proportions for mock community
  #mock_true_prop = p_mock_all %>% dplyr::select(contains("sp")),
  alr_mock_true_prop = alr_mock_true_prop,
  alr_mock_true_prop_small = alr_mock_true_prop_small,
  
  # vectors of PCR numbers
  N_pcr_samp = N_pcr_samp,
  N_pcr_mock = N_pcr_mock,
  
  # Design matrices: field samples
  N_b_samp_col = N_b_samp_col,
  model_matrix_b_samp = model_matrix_b_samp,
  model_matrix_b_samp_small = model_matrix_b_samp_small,
  model_vector_a_samp = model_vector_a_samp,
  model_vector_a_samp_small = as.array(model_vector_a_samp_small),
  
  # Design matrices: mock community samples
  N_b_mock_col = N_b_mock_col,
  model_matrix_b_mock = model_matrix_b_mock,
  model_matrix_b_mock_small = model_matrix_b_mock_small,
  model_vector_a_mock = model_vector_a_mock,
  model_vector_a_mock_small = model_vector_a_mock_small,
  
  # Priors
  alpha_prior = c(0,0.1),  # normal prior
  beta_prior = c(0,10),    # normal prior
  tau_prior = c(1.5,1.5)   # gamma prior
)


stan_pars <- c(
  #"alpha",
  "beta",
  "eta_samp",
  #"eta_mock",
  "tau",
  "mu_samp",
  #"mu_mock",
  "int_samp_small"
)
```

```{r}
#########################################
#########################################
#Bayesian Estimation
N_CHAIN = 3
Warm = 1000
Iter = 1500
Treedepth = 15
Adapt_delta = 0.70

rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())

setwd("~/nwfsc-cb-Toward_Quantitative_Metabarcoding-fba0076/code/Core_Stan_models")
## my note: this would not run without having the wd set to the Core_Stan_models folder... 

stanMod = stan(file = "quant_metabar_no_mock_no_alpha.stan" ,data = stan_data,
               verbose = FALSE, chains = N_CHAIN, thin = 1,
               warmup = Warm, iter = Warm + Iter,
               control = list(adapt_init_buffer = 175,
                              max_treedepth=Treedepth,
                              stepsize=0.01,
                              adapt_delta=Adapt_delta,
                              metric="diag_e"),
               pars = stan_pars,
               refresh = 10,
               boost_lib = NULL,
               #init = stan_init_f2(n.chain=N_CHAIN,N_species=N_species),
               sample_file = paste0("./tmpE.csv")
)

pars <- rstan::extract(stanMod, permuted = TRUE)
samp_params <- get_sampler_params(stanMod)

stanMod_summary <- list()
stanMod_summary[["tau"]] <- summary(stanMod,pars=c("tau"))$summary
stanMod_summary[["beta"]] <- summary(stanMod,pars="beta")$summary
stanMod_summary[["eta_samp_raw"]] <- summary(stanMod,pars="eta_samp")$summary
stanMod_summary[["mu_samp"]] <- summary(stanMod,pars="mu_samp")$summary
stanMod_summary[["int_samp_small"]] <- summary(stanMod,pars="int_samp_small")$summary


Output <- list(
  mockcom_data = mockcom_data, # raw data from all observations, all communities.
  env = env,  #environmental data
  mc = mc, #mock data
  Species = unique(mc$Species),
  
  # Realizations of simulations (input data)
  p_true = p_mock,
  p_samp_all = sample_data,
  p_mock_all = mock_data,
  
  # stan input objects
  stan_data = stan_data,
  Warm=Warm,
  Iter=Iter,
  
  # Fitted Objects
  stanMod = stanMod, # Full stan Model fitted object
  pars = pars, # MCMC output
  samp_params=samp_params, # Sampler information
  stanMod_summary = stanMod_summary # posterior summaries.
)
```

```{r}
save(Output,file=paste0("/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/raw_20250628_field",".Rdata"))
```


## now run stan models that adjust for amplification efficiency 

  #DESIGN MATRICES
  # mock communities first
  # species compositions (betas)
  # use mock_data  
```{r}
  N_pcr_mock <- mock_data$Cycles
  
  if(length(unique(mock_data$community))==1){
    formula_b <- Cycles ~ 1  # what is on the left side of the equation doesn't matter.
  } else {
    formula_b <- Cycles ~ community # what is on the left side of the equation doesn't matter.
  }
  model_frame <- model.frame(formula_b, mock_data)
  model_matrix_b_mock <- model.matrix(formula_b, model_frame)
  
  #formula_b <- obs_sp_1 ~ community 
  model_frame <- model.frame(formula_b, mock_data_small)
  model_matrix_b_mock_small <- model.matrix(formula_b, model_frame)
  
  # efficiencies (alphas)
  formula_a <- community ~ Cycles -1
  model_frame <- model.frame(formula_a, mock_data)
  model_vector_a_mock <- model.matrix(formula_a, model_frame) %>% as.numeric()
  model_frame <- model.frame(formula_a, mock_data_small)
  model_vector_a_mock_small <- model.matrix(formula_a, model_frame) %>% as.numeric()
  
  N_obs_mock_small <- nrow(model_matrix_b_mock_small)
  N_obs_mock       <- nrow(mock_data)
  N_b_mock_col     <- ncol(model_matrix_b_mock)  
  
  # unknown communities second
  # species compositions (betas)

  # use sample_data
  
  N_pcr_samp <- sample_data$Cycles
  
  if(length(unique(sample_data$community))==1){
    formula_b <- Cycles ~ 1  
  } else {
    formula_b <- Cycles ~ community
  }
  model_frame <- model.frame(formula_b, sample_data)
  model_matrix_b_samp <- model.matrix(formula_b, model_frame)
  
  #formula_b <- obs_sp_1 ~ community 
  #p_samp_all$site <- as.factor(p_samp_all$site)
  model_frame <- model.frame(formula_b, sample_data_small)
  model_matrix_b_samp_small <- model.matrix(formula_b, model_frame)
  
  # efficiencies (alpha)
  formula_a <- community ~ Cycles -1
  model_frame <- model.frame(formula_a, sample_data)
  model_vector_a_samp <- model.matrix(formula_a, model_frame) %>% as.numeric()
  model_frame <- model.frame(formula_a, sample_data_small)
  model_vector_a_samp_small <- model.matrix(formula_a, model_frame) %>% as.numeric()
  
  #counters 
  N_obs_samp_small <- nrow(model_matrix_b_samp_small)
  N_obs_samp <- nrow(sample_data)
  N_b_samp_col <- ncol(model_matrix_b_samp)  
  

stan_data <- list(
  N_species = N_species,   # Number of species in data
  N_obs_samp = N_obs_samp, # Number of observed samples 
  N_obs_mock = N_obs_mock, # Number of observed mock samples
  N_obs_samp_small = N_obs_samp_small, # Number of observed samples 
  N_obs_mock_small = N_obs_mock_small, # Number of observed mock samples
  
  # Observed data of community matrices
  sample_data = sample_data %>% select(-community,-Cycles,-tech_rep),
  mock_data   = mock_data  %>% select(-community,-Cycles,-tech_rep),
  
  # True proportions for mock community
  #mock_true_prop = p_mock_all %>% dplyr::select(contains("sp")),
  alr_mock_true_prop = alr_mock_true_prop,
  alr_mock_true_prop_small = alr_mock_true_prop_small,
  
  # vectors of PCR numbers
  N_pcr_samp = N_pcr_samp,
  N_pcr_mock = N_pcr_mock,
  
  # Design matrices: field samples
  N_b_samp_col = N_b_samp_col,
  model_matrix_b_samp = model_matrix_b_samp,
  model_matrix_b_samp_small = model_matrix_b_samp_small,
  model_vector_a_samp = model_vector_a_samp,
  model_vector_a_samp_small = as.array(model_vector_a_samp_small),
  
  # Design matrices: mock community samples
  N_b_mock_col = N_b_mock_col,
  model_matrix_b_mock = model_matrix_b_mock,
  model_matrix_b_mock_small = model_matrix_b_mock_small,
  model_vector_a_mock = model_vector_a_mock,
  model_vector_a_mock_small = model_vector_a_mock_small,
  
  # Priors
  alpha_prior = c(0,0.1),  # normal prior
  beta_prior = c(0,10),    # normal prior
  tau_prior = c(1.5,1.5)   # gamma prior
)


stan_pars <- c(
  "alpha",
  "beta",
  "eta_samp",
  "eta_mock",
  "tau",
  "mu_samp",
  "mu_mock",
  "int_samp_small"
)

stan_init_f2 <- function(n.chain,N_species){#J_seb,K_seb){
  A <- list()
  for(i in 1:n.chain){
    A[[i]] <- list(
      # tau = runif(N_species-1,0.1,0.5),
      alpha_raw = runif(N_species-1,-0.5,0.5)
    )
  }
  return(A)
}
```

# These are the parameters that are to be monitored during optimization or MCMC:
stan_pars <- c(
  "alpha", # efficiencies relative to the reference species
  "beta",  # parameters for each site (NOT )
  "eta_samp", # overdispersion random effects for each species-site combination (field samples)
  "eta_mock", # overdispersion random effects for each species-site combination (mock samples)
  "tau", # sd of random effects (one for each species (less the reference species))
  "mu_samp", # Predicted proportions for each species-site (unknown samples)
  "mu_mock", # Predicted proportions for each species-site (mock samples)  
  "int_samp_small" # this is the predicted intercept for each site 
)


```{r}
#########################################
#########################################
#Bayesian Estimation
N_CHAIN = 3
Warm = 1000
Iter = 1500
Treedepth = 15
Adapt_delta = 0.70

rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())

## have to change wd here
setwd("~/nwfsc-cb-Toward_Quantitative_Metabarcoding-fba0076/code/Core_Stan_models")

stanMod = stan(file = "quant_metabar_multinom.stan" ,data = stan_data,
               verbose = FALSE, chains = N_CHAIN, thin = 1,
               warmup = Warm, iter = Warm + Iter,
               control = list(adapt_init_buffer = 175,
                              max_treedepth=Treedepth,
                              stepsize=0.01,
                              adapt_delta=Adapt_delta,
                              metric="diag_e"),
               pars = stan_pars,
               refresh = 10,
               boost_lib = NULL,
               init = stan_init_f2(n.chain=N_CHAIN,N_species=N_species),
               sample_file = paste0("./tmpF.csv")
)

pars <- rstan::extract(stanMod, permuted = TRUE)
samp_params <- get_sampler_params(stanMod)

stanMod_summary <- list()
stanMod_summary[["alpha"]] <- summary(stanMod,pars="alpha")$summary
stanMod_summary[["tau"]] <- summary(stanMod,pars=c("tau"))$summary
stanMod_summary[["beta"]] <- summary(stanMod,pars="beta")$summary
stanMod_summary[["eta_samp_raw"]] <- summary(stanMod,pars="eta_samp")$summary
stanMod_summary[["eta_mock_raw"]] <- summary(stanMod,pars="eta_mock")$summary
stanMod_summary[["mu_samp"]] <- summary(stanMod,pars="mu_samp")$summary
stanMod_summary[["mu_mock"]] <- summary(stanMod,pars="mu_mock")$summary
stanMod_summary[["int_samp_small"]] <- summary(stanMod,pars="int_samp_small")$summary

Output <- list(
  mockcom_data = mockcom_data, # raw data from all observations, all communities.
  env = env,  #environmental data
  mc = mc, #mock data
  Species = unique(mc$Species),
  
  # Realizations of simulations (input data)
  p_true = p_mock,
  p_samp_all = sample_data,
  p_mock_all = mock_data,
  
  # stan input objects
  stan_data = stan_data,
  Warm=Warm,
  Iter=Iter,
  
  # Fitted Objects
  stanMod = stanMod, # Full stan Model fitted object
  pars = pars, # MCMC output
  samp_params=samp_params, # Sampler information
  stanMod_summary = stanMod_summary # posterior summaries.
)
```

```{r}
save(Output,file=paste0("/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/mock_20250618_field",".Rdata"))
```


### make figure 5 

```{r}
load("/home/kimberly.ledger/rockfish_mb/outputs/mock_quant/mock_20250618_field.Rdata")
mock1 <- Output
```

```{r}
# summarize raw estimates from reads for each species.
mock1.raw <- mock1$env %>% group_by(community,Cycles,tech_rep) %>%
      mutate(sum.ng = sum(start_conc_ng),
             true.prop = start_conc_ng / sum.ng) %>%
      ungroup() %>%
      group_by(Species,community,Cycles,true.prop) %>%
  summarize(simple.Mean=mean(propReads),
            simple.N = length(tech_rep)) %>%
  replace_na(list(raw.Mean=0,raw.SD=0,raw.SE=0))

# extract predicted proportions from the posterior
COM <- data.frame(community = levels(mock1$env$community %>% as.factor()))
COM$comm_idx <- 1:nrow(COM)
SP  <- mock1$env %>% distinct(Species,sp_idx) %>% as.data.frame()

# These are the predicted intercepts for the posteriors
beta_posterior <- mock1$stanMod_summary[["int_samp_small"]][, c(1,4:8)]
colnames(beta_posterior) <- paste0("mock1.",substr(colnames(beta_posterior),1,nchar(colnames(beta_posterior))-1))
colnames(beta_posterior)[1] <- "mock1.mean"
beta_posterior <- as.data.frame(beta_posterior)

mock1.post <-expand.grid(comm_idx = COM$comm_idx,sp_idx =SP$sp_idx) %>% 
    arrange(comm_idx,sp_idx) %>% 
    left_join(.,COM) %>% 
    left_join(.,SP) %>% 
    bind_cols(.,beta_posterior)

# Combine the raw estimates and posterior estimates
mock1.all <- full_join(mock1.raw,mock1.post)
```

filter to just have the species actually present in the survey and non-zero reads in simple.Mean
```{r}
mock1.all.filtered <- mock1.all %>%
  filter(!is.na(simple.Mean)) %>%
  mutate(taxon = ifelse(Species == "zRefSpecies_Sebastes alutus", "Sebastes alutus", Species)) %>%
  mutate(taxon = ifelse(Species == "Sebastes ciliatus", "Sebastes ciliatus_variabilis", taxon)) %>%
  mutate(taxon = ifelse(community == "e02312" & Species == "Sebastes ciliatus", "Sebastes ciliatus_variabilis_polyspinus", taxon))
```

reread in the metadata 
```{r}
field_table_w_meta <- read.csv("/home/kimberly.ledger/rockfish_mb/outputs/field_quant/raw_field_data.csv") %>%
  select(extraction_ID, station, taxon) %>%
  rename(community = extraction_ID)
```

join bias-corrected proportions to the metadata 
```{r}
corrected <- field_table_w_meta %>%
  left_join(mock1.all.filtered) %>%
  mutate(taxon_abbr = sub("^(\\w)\\w*\\s(\\w+)", "\\1. \\2", taxon))
```


```{r}
my_colors <- c("#f5a056",  "#6e9bf4", "#cc7af4", "#1f77b4", "#77ab31")

bt_plot2 <- corrected %>%
  mutate(station_dummy = dense_rank(station)) %>%
  ggplot(aes(x=community, y=mock1.mean, fill=taxon_abbr)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  scale_fill_manual(values = my_colors, na.translate = FALSE) + # Removes NA from the legend
  facet_grid(~station_dummy, scales = "free_x") +
  labs(
    y = "bias corrected read abundance (%)",
    x = "samples by haul") + 
  theme_bw() +
  theme(axis.text.x = element_blank(),
    #axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8, face = "italic"),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "bottom",
    legend.title = element_blank(),
    #strip.text = element_blank(), # Removes facet labels
    axis.ticks.x = element_blank() 
  )
  
bt_plot2
```

```{r}
ggsave("/home/kimberly.ledger/rockfish_mb/figures/bt2_w_zeros_dummyfacets_biascorrected.png", plot = bt_plot2, dpi = 300, width = 8, height =3)
```
